refactor(web): migrate types/* to ReScript (migration Phase 1)
Port the 10 plain record DTOs in src/types to ReScript records exposed
via genType. Each `Foo.res` defines `@genType type t = {...}`; genType
emits Foo.gen.tsx with the identical TS shape, and the 11 consumer files
switch their import to `{ t as Foo } from "../types/Foo.gen"` (local usage
names unchanged).

Faithful mappings:
- number -> float (genType renders both as `number`; float is the honest
  shape for JSON-sourced values).
- `string[] | null` -> Null.t<array<string>>, `T | null` -> Null.t<T>.
- optional fields (`albumArt?`, `description?`) -> ReScript optional fields.
- AccessToken's `& { token }` intersection -> a spelled-out `created` record.
- Profile's repeated cloud-drive object -> a shared `cloudDrive` record.

search.ts is intentionally left in TS: it carries runtime type guards and
`Federation<T>` generics/discriminated unions that don't map cleanly yet
(revisit alongside the API layer in Phase 3).

Verified: rescript build, tsc -b, and full vite build all pass.
