+++ apps/web/src/types/Album.res
@@ -1,0 +1,14 @@
+@genType
+type t = {
+  id: string,
+  uri: string,
+  title: string,
+  artist: string,
+  artistUri: string,
+  year: float,
+  albumArt: string,
+  releaseDate: string,
+  sha256: string,
+  playCount: float,
+  uniqueListeners: float,
+}
