+++ apps/web/src/types/ApiKey.res
@@ -1,0 +1,10 @@
+@genType
+type t = {
+  id: string,
+  name: string,
+  description?: string,
+  apiKey: string,
+  sharedSecret: string,
+  enabled: bool,
+  createdAt: string,
+}
