+++ apps/web/src/types/Artist.res
@@ -1,0 +1,11 @@
+@genType
+type t = {
+  id: string,
+  name: string,
+  picture: string,
+  playCount: float,
+  sha256: string,
+  tags: Null.t<array<string>>,
+  uniqueListeners: float,
+  uri: string,
+}
