+++ apps/web/src/types/File.res
@@ -1,0 +1,6 @@
+@genType
+type t = {
+  id: string,
+  name: string,
+  tag: string,
+}
