+++ apps/web/src/types/Scrobble.res
@@ -1,0 +1,16 @@
+@genType
+type t = {
+  id: string,
+  trackId: string,
+  title: string,
+  artist: string,
+  album: string,
+  albumArt?: string,
+  albumArtist: string,
+  handle: string,
+  trackUri: string,
+  albumUri: string,
+  artistUri: string,
+  uri: string,
+  createdAt: string,
+}
