--- apps/web/src/types/access-token.ts
@@ -1,12 +1,0 @@
-export type AccessToken = {
-  id: string;
-  name: string;
-  lastFour: string;
-  lastUsedAt: string | null;
-  createdAt: string;
-  updatedAt: string;
-};
-
-export type CreatedAccessToken = AccessToken & {
-  token: string;
-};
