--- apps/web/src/types/album.ts
@@ -1,13 +1,0 @@
-export type Album = {
-  id: string;
-  uri: string;
-  title: string;
-  artist: string;
-  artistUri: string;
-  year: number;
-  albumArt: string;
-  releaseDate: string;
-  sha256: string;
-  playCount: number;
-  uniqueListeners: number;
-};
