--- apps/web/src/types/apikey.ts
@@ -1,9 +1,0 @@
-export type ApiKey = {
-  id: string;
-  name: string;
-  description?: string;
-  apiKey: string;
-  sharedSecret: string;
-  enabled: boolean;
-  createdAt: string;
-};
