--- apps/web/src/types/file.ts
@@ -1,5 +1,0 @@
-export type File = {
-  id: string;
-  name: string;
-  tag: string;
-};
