--- apps/web/src/types/track.ts
@@ -1,19 +1,0 @@
-export type Track = {
-  id: string;
-  uri: string;
-  uniqueListeners: number;
-  playCount: number;
-  title: string;
-  artist: string;
-  artistUri: string;
-  album: string;
-  albumUri: string;
-  albumArt: string;
-  albumArtist: string;
-  copyrightMessage: string;
-  discNumber: number;
-  duration: number;
-  sha256: string;
-  track_number: number;
-  created_at: string;
-};
